+++ rs/test_utilities/state/src/lib.rs
@@ -994,7 +994,7 @@ prop_compose! {
         (signals_end, reject_signals) in arb_stream_signals(
             0..=10000,
             min_signal_count..=max_signal_count,
-            with_reject_reasons
+            with_reject_reasons,
         ),
         responses_only in any::<bool>(),
     ) -> StreamHeader {
@@ -1010,6 +1010,27 @@ prop_compose! {
                 deprecated_responses_only: responses_only,
             },
         )
+    }
+}
+
+prop_compose! {
+    pub fn arb_invalid_stream_header(
+        min_signal_count: usize,
+        max_signal_count: usize,
+    )(
+        valid_stream_header in arb_stream_header(min_signal_count, max_signal_count, RejectReason::all()),
+        reason in proptest::sample::select(RejectReason::all()),
+    ) -> StreamHeader {
+        let begin = valid_stream_header.begin();
+        let end = valid_stream_header.end();
+        let signals_end = valid_stream_header.signals_end();
+        let mut reject_signals = valid_stream_header.reject_signals().clone();
+        let flags = *valid_stream_header.flags();
+
+        // `reject_signals` may not contain the `signals_end`.
+        reject_signals.push_back(RejectSignal::new(reason, signals_end));
+
+        StreamHeader::new(begin, end, signals_end, reject_signals, flags)
     }
 }